requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
rich==13.6.0
newsapi-python==0.2.7
click==8.1.7 